[Refactor]: repository 파일 이동 및 querydsl 메서드 구현
Related to: #559
