+++ src/main/java/org/sopt/makers/internal/community/repository/anonymous/AnonymousPostProfileRepository.java
@@ -8,18 +8,13 @@ import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
 import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnonymousPostProfileRepository extends JpaRepository<AnonymousPostProfile, Long> {
+public interface AnonymousPostProfileRepository extends JpaRepository<AnonymousPostProfile, Long>, AnonymousPostProfileRepositoryCustom {
 
 	// CREATE
 
 	// READ
 	Optional<AnonymousPostProfile> findAnonymousPostProfileByMemberIdAndCommunityPostId(Long memberId, Long communityPostId);
-
 	Optional<AnonymousPostProfile> findByMemberAndCommunityPost(Member member, CommunityPost post);
-
-	List<AnonymousPostProfile> findTop4ByOrderByCreatedAtDesc();
-
-	List<AnonymousPostProfile> findTop50ByOrderByCreatedAtDesc();
 
 	// UPDATE
