+++ src/main/java/org/sopt/makers/internal/community/repository/anonymous/AnonymousPostProfileRepositoryCustom.java
@@ -1,0 +1,10 @@
+package org.sopt.makers.internal.community.repository.anonymous;
+
+import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
+
+import java.util.List;
+
+public interface AnonymousPostProfileRepositoryCustom {
+
+    List<AnonymousPostProfile> findTopByOrderByIdDescWithLimit(int limit);
+}
